Primary Coding Finished
The first release is soon! It's time for heavy testing in every scenario I can think of. Maybe a few minor coding changes from here for fixes, but the primary coding is finished. I added the custom form, added all the custom data by hand, made the custom functionality similar to the rest.